% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_geography.R
\name{create_geography}
\alias{create_geography}
\title{Create geography}
\usage{
create_geography(
  date = Sys.Date(),
  subset = NULL,
  aspect_ratio = NULL,
  zoom = 0.9,
  show_non_member_states = TRUE,
  insets = NULL,
  projection = "lambert_azimuthal_equal_area",
  resolution = "high"
)
}
\arguments{
\item{date}{String or date. Required, and has a default value. The date of
the data. If a string, needs to have the format \code{YYYY-MM-DD}. The default
is today's date.}

\item{subset}{String. Optional. A vector of member state names. The map will
be centered on these member states. If \code{NULL}, the map will be centered on
all member states. Any countries in the vector that aren't member states on
the date indicated by \code{date} will be ignored. The default value is \code{NULL}.}

\item{aspect_ratio}{Numeric. Optional. The aspect ratio of the map. The
minimum value is \code{0.5} and the maximum value is \code{2}. If \code{NULL}, the aspect
ratio will be determined by the bounding box surrounding the selected
member states. The default value is \code{NULL}.}

\item{zoom}{Numeric. Required, and has a default value. The zoom factor. A
value of \code{1} focuses the map tightly around the selected member states, and
values less than \code{1} zoom out the map. The minimum value is \code{0.5} and the
maximum value is \code{1}. The default value is \code{0.9}.}

\item{show_non_member_states}{Logical. Required, and has a default value.
Whether or not to plot non-member states. The default value is \code{TRUE}.}

\item{insets}{String. Optional. A vector of member states to create insets
for. Possible values include \code{Luxembourg}, \code{Cyprus}, and \code{Malta}. Other
countries will be ignored. If \code{NULL}, no insets are created. The default
value is \code{NULL}.}

\item{projection}{String. Required, and has a default value. The map
projection to use. The possible values are \code{equidistant_conic},
\code{lambert_conformal_conic}, \code{albers_equal_area_conic},
\code{lambert_azimuthal_equal_area}, and \code{mercator}. The default value is
\code{lambert_azimuthal_equal_area}.}

\item{resolution}{String. Required, and has a default value. The resolution
to use for the border data. The possible values are \code{high} and \code{low}. The
default is \code{high}.}
}
\value{
This function returns an object of type \code{eumaps.geography}, which you
can pass to the \code{geography} argument of \code{make_maps()}.
}
\description{
This function creates the geography for a map of the European
Union (EU) made by \code{eumaps::make_map()}. It creates a object of type
\code{eumaps.geography}, which you can pass to the \code{geography} argument of
\code{make_maps()}.

\code{create_geography()} will automatically center the map on all countries
that are member states on the date indicated by the argument \code{date}. This
way, you never have to specify the bounds of the map, which can be
complicated, depending on the map projection you want to use. You also
don't have to know the accession dates of the member states. By default,
\code{date} is set to today's date. You can use the optional argument \code{subset}
to center the map on a subset of member states.

You can use \code{create_geography()} to set a variety of other options that
make it easy to make the map look exactly how you want.

First, you can use the argument \code{aspect_ratio} to specify the aspect ratio
of the map. A value greater than \code{1} makes a map that is wider than it is
tall and a value less than \code{1} makes a map that is taller than it is wide.
The minimum value is \code{0.5} and the maximum value is \code{2}.

Second, you can use \code{zoom} to choose a zoom factor. A value of \code{1} focuses
the map tightly around the selected member states, and values less than \code{1}
zoom out the map. The minimum value is \code{0.5}. The default is \code{0.9}.

Third, you can use \code{show_non_member_states} to choose whether or not to
plot non-member states. The default value is \code{TRUE}.

Fourth, you can use \code{insets} to create insets for member states whose
values would be hard to read otherwise. By default, \code{insets} equals \code{NULL},
in which case no insets are created. You can provide a vector of member
state names to create insets for those member states. You can create insets
for Luxembourg, Malta, and Cyprus. The insets will appear in the top right
corner of the map in the order given by the vector.

Fifth, you can use \code{projection} to choose between 5 common map projections
that are appropriate for Europe. You can run \code{list_projections()} to see
the possible values.

Finally, you can use \code{resolution} to choose between low or high resolution
border data. The default value is \code{high} and the alternative is \code{low}. The
map will take longer to render if you use the high resolution data. The
function always uses high resolution border data for the insets.
}
\examples{
# using the default values
geography <- create_geography()

# using all options
create_geography <- function(
  date = Sys.Date(),
  subset = NULL,
  aspect_ratio = NULL,
  zoom = 0.9,
  show_non_member_states = TRUE,
  insets = NULL,
  projection = "lambert_azimuthal_equal_area",
  resolution = "high"
)

}
