% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_member_states.R
\name{list_member_states}
\alias{list_member_states}
\title{List member states}
\usage{
list_member_states()
}
\value{
This function returns a \code{tibble} with 2 columns: \code{member_state_id}
and \code{member_state}.
}
\description{
This function lists all current and former European Union (EU)
member states. These are the values that are accepted by the \code{eumaps}
package.
}
\examples{
# get a list of all current and former member states
list_member_states()

}
