% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate map data}
\usage{
simulate_data(date = NULL, missing = NULL)
}
\arguments{
\item{date}{String or date. Optional. If not \code{NULL}, data will only be
simulated for countries that were member states on this date. If not
\code{NULL}, today's date is used, and data will only be simulated for current
member states. The default value is \code{NULL}.}

\item{missing}{String. Optional. A vector of member states names. The data
for these member states will be coded as \code{NA}. This is option is useful for
testing out color palettes. The default value is \code{NULL}.}
}
\value{
This function returns a \code{tibble} with 3 columns: \code{member_state_id},
\code{member_state}, and \code{value}.
}
\description{
This function simulates data for a map of the European Union
(EU) made by \code{eumaps::make_map()}. The data is drawn from a standard
uniform distribution. It creates a tibble. You can pass the column
\code{member_state} to the \code{member_states} argument of \code{create_palette()} and
the column \code{value} to the \code{values} argument of \code{create_palette()}.
}
\examples{
# using the default values
data <- simulate_data()

# using all options
data <- simulate_data(
  date = "2015-01-01",
  missing = c("France", "Germany")
)

}
